initial new turbine simulation, still needs some tweaking

+++ src/main/java/net/roguelogix/biggerreactors/classic/turbine/simulation/classic/ClassicTurbineSimulation.java
@@ -276,7 +276,7 @@ public class ClassicTurbineSimulation implements ITurbineSimulation {
     }
     
     @Override
-    public long rotorMass() {
+    public double rotorMass() {
         return rotorMass;
     }
